scraper/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
pydantic==2.9.2
python-dateutil==2.9.0.post0
